+++ changelog.md
@@ -3,6 +3,7 @@
 //  Add your own contribution below
 
 * Started turning this into a real project by adding tests - orta
+* `danger.pr` -> `danger.github.pr`, I've also created interfaces for them - orta
 
 ### 0.0.5-0.0.10